Have index.js just run the other scripts
